+++ chains/avalanche/Cargo.toml
@@ -22,6 +22,7 @@ sha3 = "0.10"
 tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
 tracing = "0.1.40"
 url = "2.4"
+utils = { path = "../utils" }
 
 [dev-dependencies]
 serial_test = "3.1.1"
